Allow http GETs if we are a public server.
This is determined from a command line parameter.

Also allow JS to use gets if we are not on localhost.
