+++ src/peergos/server/net/CoreNodeHandler.java
@@ -1,0 +1,147 @@
+package peergos.server.net;
+
+import com.sun.net.httpserver.*;
+import peergos.server.*;
+import peergos.server.util.*;
+import peergos.shared.cbor.*;
+import peergos.shared.corenode.*;
+import peergos.shared.crypto.hash.*;
+import peergos.shared.io.ipfs.api.*;
+import peergos.shared.util.*;
+
+import java.io.*;
+import java.util.*;
+import java.util.logging.*;
+import java.util.zip.*;
+
+public class CoreNodeHandler implements HttpHandler
+{
+    private static final Logger LOG = Logging.LOG();
+
+    private final CoreNode coreNode;
+    private final boolean isPublicServer;
+
+    public CoreNodeHandler(CoreNode coreNode, boolean isPublicServer) {
+        this.coreNode = coreNode;
+        this.isPublicServer = isPublicServer;
+    }
+
+    public void handle(HttpExchange exchange)
+    {
+        long t1 = System.currentTimeMillis();
+        DataInputStream din = new DataInputStream(exchange.getRequestBody());
+
+        ByteArrayOutputStream bout = new ByteArrayOutputStream();
+        DataOutputStream dout = new DataOutputStream(bout);
+
+        String path = exchange.getRequestURI().getPath();
+        if (path.startsWith("/"))
+            path = path.substring(1);
+        String[] subComponents = path.substring(Constants.CORE_URL.length()).split("/");
+        String method = subComponents[0];
+
+        try {
+            if (! HttpUtil.allowedQuery(exchange, isPublicServer)) {
+                exchange.sendResponseHeaders(405, 0);
+                return;
+            }
+
+            switch (method)
+            {
+                case "getChain":
+                    AggregatedMetrics.GET_PUBLIC_KEY_CHAIN.inc();
+                    getChain(din, dout);
+                    break;
+                case "updateChain":
+                    AggregatedMetrics.UPDATE_PUBLIC_KEY_CHAIN.inc();
+                    updateChain(din, dout);
+                    break;
+                case "getPublicKey":
+                    AggregatedMetrics.GET_PUBLIC_KEY.inc();
+                    getPublicKey(din, dout);
+                    break;
+                case "getUsername":
+                    AggregatedMetrics.GET_USERNAME.inc();
+                    getUsername(din, dout);
+                    break;
+                case "getUsernamesGzip":
+                    AggregatedMetrics.GET_ALL_USERNAMES.inc();
+                    exchange.getResponseHeaders().set("Content-Encoding", "gzip");
+                    exchange.getResponseHeaders().set("Content-Type", "application/json");
+                    getAllUsernamesGzip(subComponents.length > 1 ? subComponents[1] : "", din, dout);
+                    break;
+                default:
+                    throw new IOException("Unknown pkinode method!");
+            }
+
+            dout.flush();
+            dout.close();
+            byte[] b = bout.toByteArray();
+            exchange.sendResponseHeaders(200, b.length);
+            exchange.getResponseBody().write(b);
+        } catch (Exception e) {
+            HttpUtil.replyError(exchange, e);
+        } finally {
+            exchange.close();
+            long t2 = System.currentTimeMillis();
+            LOG.info("Corenode server handled " + method + " request in: " + (t2 - t1) + " mS");
+        }
+    }
+
+    void getChain(DataInputStream din, DataOutputStream dout) throws Exception
+    {
+        String username = CoreNodeUtils.deserializeString(din);
+
+        List<UserPublicKeyLink> chain = coreNode.getChain(username).get();
+        dout.write(new CborObject.CborList(chain).serialize());
+    }
+
+    void updateChain(DataInputStream din, DataOutputStream dout) throws Exception
+    {
+        String username = CoreNodeUtils.deserializeString(din);
+        int count = din.readInt();
+        List<UserPublicKeyLink> res = new ArrayList<>();
+        for (int i=0; i < count; i++) {
+            res.add(UserPublicKeyLink.fromCbor(CborObject.fromByteArray(Serialize.deserializeByteArray(din, UserPublicKeyLink.MAX_SIZE))));
+        }
+        boolean isAdded = coreNode.updateChain(username, res).get();
+
+        dout.writeBoolean(isAdded);
+    }
+
+    void getPublicKey(DataInputStream din, DataOutputStream dout) throws Exception
+    {
+        String username = CoreNodeUtils.deserializeString(din);
+        Optional<PublicKeyHash> k = coreNode.getPublicKeyHash(username).get();
+        dout.writeBoolean(k.isPresent());
+        if (!k.isPresent())
+            return;
+        byte[] b = k.get().serialize();
+        dout.writeInt(b.length);
+        dout.write(b);
+    }
+
+    void getUsername(DataInputStream din, DataOutputStream dout) throws Exception
+    {
+        byte[] publicKey = CoreNodeUtils.deserializeByteArray(din);
+        String k = coreNode.getUsername(PublicKeyHash.fromCbor(CborObject.fromByteArray(publicKey))).get();
+        if (k == null)
+            k="";
+        Serialize.serialize(k, dout);
+    }
+
+    void getAllUsernamesGzip(String prefix, DataInputStream din, DataOutputStream dout) throws Exception
+    {
+        List<String> res = coreNode.getUsernames(prefix).get();
+        ByteArrayOutputStream bout = new ByteArrayOutputStream();
+        GZIPOutputStream gout = new GZIPOutputStream(bout);
+        gout.write(JSONParser.toString(res).getBytes());
+        gout.flush();
+        gout.close();
+        dout.write(bout.toByteArray());
+    }
+
+    public void close() throws IOException{
+        coreNode.close();
+    }
+}
