+++ src/peergos/server/net/DHTHandler.java
@@ -29,21 +29,25 @@ public class DHTHandler implements HttpHandler {
     private final Hasher hasher;
     private final BiFunction<PublicKeyHash, Integer, Boolean> keyFilter;
     private final String apiPrefix;
+    private final boolean isPublicServer;
 
     public DHTHandler(ContentAddressedStorage dht,
                       Hasher hasher,
                       BiFunction<PublicKeyHash, Integer, Boolean> keyFilter,
-                      String apiPrefix) {
+                      String apiPrefix,
+                      boolean isPublicServer) {
         this.dht = dht;
         this.hasher = hasher;
         this.keyFilter = keyFilter;
         this.apiPrefix = apiPrefix;
+        this.isPublicServer = isPublicServer;
     }
 
     public DHTHandler(ContentAddressedStorage dht,
                       Hasher hasher,
-                      BiFunction<PublicKeyHash, Integer, Boolean> keyFilter) {
-        this(dht, hasher, keyFilter, "/api/v0/");
+                      BiFunction<PublicKeyHash, Integer, Boolean> keyFilter,
+                      boolean isPublicServer) {
+        this(dht, hasher, keyFilter, "/api/v0/", isPublicServer);
     }
 
     @Override
@@ -51,8 +55,7 @@ public class DHTHandler implements HttpHandler {
         long t1 = System.currentTimeMillis();
         String path = httpExchange.getRequestURI().getPath();
         try {
-            // only allow http POST requests
-            if (! httpExchange.getRequestMethod().equals("POST")) {
+            if (! HttpUtil.allowedQuery(httpExchange, isPublicServer)) {
                 httpExchange.sendResponseHeaders(405, 0);
                 return;
             }
