+++ src/peergos/server/net/SpaceHandler.java
@@ -16,13 +16,15 @@ import java.util.logging.*;
 /** This is the http endpoint for SpaceUsage calls
  *
  */
-public class StorageHandler implements HttpHandler {
+public class SpaceHandler implements HttpHandler {
     private static final Logger LOG = Logging.LOG();
 
     private final SpaceUsage spaceUsage;
+    private final boolean isPublicServer;
 
-    public StorageHandler(SpaceUsage spaceUsage) {
+    public SpaceHandler(SpaceUsage spaceUsage, boolean isPublicServer) {
         this.spaceUsage = spaceUsage;
+        this.isPublicServer = isPublicServer;
     }
 
     public void handle(HttpExchange exchange)
@@ -40,8 +42,7 @@ public class StorageHandler implements HttpHandler {
 
         Cborable result;
         try {
-            // only allow http POST requests
-            if (!exchange.getRequestMethod().equals("POST")) {
+            if (! HttpUtil.allowedQuery(exchange, isPublicServer)) {
                 exchange.sendResponseHeaders(405, 0);
                 return;
             }
